api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg[binary]==3.2.3
pydantic==2.10.2
aio-pika==9.5.3
python-dotenv==1.0.1
